refactor: combine 1d/2d/3d plotters  into one
