+++ massdash/server/FeatureMapPlotterServer.py
@@ -1,29 +1,30 @@
 """
-massdash/server/OneDimensionPlotterServer
+massdash/server/FeatureMapPlotterServer
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 """
 
 from typing import Dict
 from os.path import basename
+from bokeh.layouts import gridplot
 from collections import defaultdict
 import streamlit as st
 
 # Loaders
 from ..loaders.GenericChromatogramLoader import GenericChromatogramLoader
 # Plotting
-from ..plotting import PlotConfig, InteractivePlotter
+from ..plotting import PlotConfig, InteractivePlotter, InteractiveThreeDimensionPlotter, InteractiveTwoDimensionPlotter
 # Server
 from .PeakPickingServer import PeakPickingServer
 # Structs
-from ..structs import TransitionGroup, TransitionGroupFeature, FeatureMapCollection, FeatureMap, TransitionGroupCollection, TransitionGroupFeatureCollection
+from ..structs import FeatureMapCollection, TransitionGroupCollection, TransitionGroupFeatureCollection
 # UI
 from ..ui.TransitionListUISettings import TransitionListUISettings
 from ..ui.ChromatogramPlotUISettings import ChromatogramPlotUISettings
 from ..ui.PeakPickingUISettings import PeakPickingUISettings
 
-class OneDimensionPlotterServer:
+class FeatureMapPlotterServer:
     """
-    A class that generates one dimensional plots from ChromatogramLoader data
+    A class that generates plots from FeatureMaps
 
     Args:
         feature_map_dict (dict): A dictionary containing transition groups.
@@ -58,7 +59,26 @@ class OneDimensionPlotterServer:
         self.plot_obj_dict = defaultdict(list) # a dictionary of plot objects with the mapping <RunName>:[<PlotObject>] (multiple plots per run allowed)
         self.verbose = verbose
 
+
     def generate_plots(self):
+        """
+        Generates plots based on user settings
+
+        Returns:
+            None
+        """
+        if self.chrom_plot_settings.display_plot_dimension_type == '1D':
+            self.generate_one_dimensional_plots()
+        elif self.chrom_plot_settings.display_plot_dimension_type == '2D':
+            self.generate_two_dimensional_plots()
+        elif self.chrom_plot_settings.display_plot_dimension_type == '3D':
+            self.generate_three_dimensional_plots()
+        else:
+            raise ValueError(f"Invalid plot dimension: {self.chrom_plot_settings.plot_dimension}. Valid options are '1D', '2D', and '3D'.")
+
+        return self
+
+    def generate_one_dimensional_plots(self):
         """
         Generates plots for each file in the transition group dictionary.
         """
@@ -77,7 +97,7 @@ class OneDimensionPlotterServer:
                 tr_group_feature_data = self.chrom_loader.loadTransitionGroupFeatures(self.transition_list_ui.transition_settings.selected_peptide,
                     self.transition_list_ui.transition_settings.selected_charge)
 
-            elif self.peak_picking_settings.do_peak_picking in ['pyPeakPickerMRM', 'MRMTransitionGroupPicker', 'ConformerPeakPicker']:
+            elif self.peak_picking_settings.do_peak_picking in ['pyPeakPickerMRM', 'MRMTransitionGroupPicker', 'Conformer']:
                 # Perform peak picking if enabled
                 peak_picker = PeakPickingServer(self.peak_picking_settings, self.chrom_plot_settings)
                 tr_group_feature_data = peak_picker.perform_peak_picking(tr_group_data=chromatograms, 
@@ -87,7 +107,7 @@ class OneDimensionPlotterServer:
                 pass
 
             else:
-                raise ValueError(f"Invalid peak picking algorithm: {self.peak_picking_settings.do_peak_picking}. Valid options are 'Feature File Boundaries', 'pyPeakPickerMRM', 'MRMTransitionGroupPicker', and 'ConformerPeakPicker'.")
+                raise ValueError(f"Invalid peak picking algorithm: {self.peak_picking_settings.do_peak_picking}. Valid options are 'Feature File Boundaries', 'pyPeakPickerMRM', 'MRMTransitionGroupPicker', and 'Conformer'.")
 
             plot_settings_dict = self._get_plot_settings('Retention Time (s)', 'Intensity', 'chromatogram')
             self._generate_plots_helper(chromatograms, plot_settings_dict, tr_group_feature_data)
@@ -96,7 +116,46 @@ class OneDimensionPlotterServer:
             mobilograms = self.featureMapCollection.to_mobilograms()
             plot_settings_dict = self._get_plot_settings('Ion Mobility (1/K0)', 'Intensity', 'mobilogram')
             self._generate_plots_helper(mobilograms, plot_settings_dict)
-        
+
+    def generate_two_dimensional_plots(self):
+        """
+        Generates two-dimensional plots based on the targeted data.
+
+        For each file in the targeted data, if the data is not empty, it retrieves the plot settings,
+        creates an instance of InteractiveTwoDimensionPlotter, generates the two-dimensional plots,
+        and stores them in the plot_obj_dict.
+
+        Returns:
+            None
+        """
+        for file, tr_df in self.targeted_data.items():
+            if not tr_df.empty():
+                plot_settings_dict = self._get_plot_settings(file)
+                plot_config = PlotConfig()
+                plot_config.update(plot_settings_dict)
+                plotter = InteractiveTwoDimensionPlotter(plot_config)
+                two_d_plots = plotter.plot(tr_df)
+                p = gridplot(two_d_plots, ncols=self.chrom_plot_settings.num_plot_columns, sizing_mode='stretch_width')
+                self.plot_obj_dict[file] = p
+        return self
+    
+    def generate_three_dimensional_plots(self):
+        """
+        Generates three-dimensional plots based on the targeted data.
+
+        Returns:
+            None
+        """
+        for file, tr_df in self.targeted_data.items():
+            if not tr_df.empty():
+                plot_settings_dict = self._get_plot_settings(file)
+                plot_config = PlotConfig()
+                plot_config.update(plot_settings_dict)
+                plotter = InteractiveThreeDimensionPlotter(plot_config)
+                three_d_plots = plotter.plot(tr_df)
+                self.plot_obj_dict[file] = three_d_plots
+        return self
+    
     def _get_plot_settings(self, x_label: str, y_label: str, plot_type=None) -> dict:
         """
         Returns the plot settings dictionary for a given plot type.
@@ -144,3 +203,4 @@ class OneDimensionPlotterServer:
                                                             plot_type=plot_settings_dict['plot_type']))
         except ValueError:
             st.error("Failed to generate plot! There may be no data for selected transition group.")
+
