+++ massdash/server/__init__.py
@@ -6,14 +6,10 @@ This subpackage contains "server side" classes for GUI
 """
 from .ExtractedIonChromatogramAnalysisServer import ExtractedIonChromatogramAnalysisServer
 from .PeakPickingServer import PeakPickingServer
-from .OneDimensionPlotterServer import OneDimensionPlotterServer
+from .FeatureMapPlotterServer import FeatureMapPlotterServer
 from .RawTargetedExtractionAnalysisServer import RawTargetedExtractionAnalysisServer
-from .ThreeDimensionPlotterServer import ThreeDimensionPlotterServer
-from .TwoDimensionPlotterServer import TwoDimensionPlotterServer
 
 __all__ = [ "ExtractedIonChromatogramAnalysisServer", 
             "PeakPickingServer", 
-            "OneDimensionPlotterServer", 
-            "RawTargetedExtractionAnalysisServer", 
-            "ThreeDimensionPlotterServer", 
-            "TwoDimensionPlotterServer" ]
+            "FeatureMapPlotterServer", 
+            "RawTargetedExtractionAnalysisServer"]
